feat: Basic implementation for matching listings
Currently, the following tables exist:

  * `users`
  * `listings`
  * `transactions`
  * `items`

With these, we're able to create automatically match listings, and
generate transactions as metadata. This matches based on price but it is
a rather naive implementation since it isn't well-optimized, neither does
it attempt to fulfill the entire quantity of the listing. Rather, it just
matches it against the matching listing with the highest quantity at the
same price. A matching listing of a BUY listing is a SELL listing, and
vice versa.

Further improvements can be made:

  * Improve search performance by utilizing indices
  * Usage of row level policies to handle authorization
  * Fulfill the entire listing by searching for the necessary amount of
    listings rather than just finding one.
  * Automatically close the listing after a certain period of time since
    the last transaction.

Once these have been completed, it should be possible to build the rest
of the system (e.g back-end & front-end).