Add documentation for project
I'll be using `ex_doc` for the documentation. It's not necessary to
use this, in fact I think it'll be better to use something like Sphinx,
and ReadTheDocs, but I prefer the UI of `ex_doc`.

So this will contain the documentation for the entire project, or at
least in terms of how to use it.