refactor: Match to potentially many listings
`transact/0` looks for matching listings that follow a certain
criteria. This is going replace the old approach of looking for matches.
The advantage of this one is that it can match with multiple listings
to better fulfill the quota required; something approach couldn't do,
since it could only find one matching listing.