feat: Session-based authentication (#33)
I did not want to deal with JWTs for handling user sessions since it
seemed like the wrong tool for the job
(http://cryto.net/~joepie91/blog/2016/06/13/stop-using-jwt-for-sessions/).
Unfortunately, PostgREST comes with JWTs as the default. So this
required a lot of boilerplate to get sessions via cookies working.

With that aside, there are a couple new session-related endpoints:

  * `/rpc/login`
  * `/rpc/register`
  * `/rpc/refresh_session`
  * `/rpc/logout`

These haven't been documented yet, but will be in the future updates,
along with which tables/stored procedures are accessible with a given
role.