refactor: Reorganize migration scripts (#34)
Scripts are (somewhat) better to read now. Each script applicable is
prefixed with the schema they belong to. Permissions and comments are
also declared at the same place as the object's creation.

This is all untested so far though. But based on my limited testing with
Insomnia, it _seems_ good. I have to write tests with `pgTAP` for some
sanity check.