feat: Integrate batching (#35)
Matching new listings used to disregard `batch`, but now it doesn't. To
make dealing with batches better, it normalizes the quantity, cost, and
batch. The reason is there may be two listings that are essentially the
same thing, but computed differently.

Here is an example:

**LISTING A**
quantity: 20
cost: 20
batch: 2

**LISTING B**
quantity: 40
cost: 10
batch: 1

These two listings are identical. It is only because of it being batched
differently that makes it _appear_ different. So now it doesn't matter
if you encode it as **A** or **B**, GNAWEX will normalize it, and it
will all be like **B** once recorded in the system.